New Features 10/21/2020
Added a navigation drawer menu to Home screen and Manager Controls screen.
Added more buttons for home screen for managers and employees.
Added a scroll view to display current schedule and allow enough room to display buttons.
Displays the 'manager controls button' only for managers, otherwise hides it.
Ability to go back to the home screen or logout using the navigation menu.
Reorganization of files.
Displays a toast message when the user successfully logs in or not.
Other minor changes.