Login Screen API POST
Allows users to login,
Created a new LoginObject which contains login information as well as variables for responses.
Added the api post request for login in ApiCalls.kt
Receives an auth_token in response which is stored on a variable named auth_token which can be used throughout the application.
(Kept the old local users in the code for testing purposes)