Allows users to login
Allows users to login,
Created a new LoginObject which contains login information as well as variables for responses.
Added the api post request for login in ApiCalls.kt
Receives an auth_token in response which is stored on a variable named auth_token which can be used throughout the application.
Users are now directed back to the home screen after succesfull registration
(Displays the authorization token as a TOAST after login just for testing purposes to ensure its working. Will be removed later)
(Kept the old local users in the code for testing purposes)