Logout API implementation AND auth_token save
Added API calls for logging out the user.
Added code so that once the user is logged out, they can no longer go back to the previous screen unless they log in.
The auth_token is now saved in all the activities for futher use.
Created a new file for retrofit builder so the code is easier to maintain.