Request Time off activity and xml
Created the request time off activity file and xml file.
Added file to manifest.
Created a new layout for the file.
Added code to make the buttons functional on the activity and on the previos activity.
Saved the token_code in the new activity to allow communication with the server.