Create new shift screen
Added a .kt file and a xml file for creating new shift.
Added functions to buttons for access to and from the new activtiy.
Added the new activity in manifest.
Added a spinner to display a list of employees to select along with date and time of shift.
Added temporary employee names that will be replaced with data recieved from GET request later.