Approve Time off request activity
Created a new activity for managers to approve time off requests.
Created the xml layout file for the new time off request activtity.
Added button functions for the new activity and any activities that connect to the new one.
Added activitity in android manifest.
Changed the layout for the "request time off activity for employees"