Drop Shift API call
Allows the users to see a list of assigned shifts and click on the drop button to drop the shift.
Modified the adapter to include retrofit code for sending drop request if the image button is clicked.
Created a DropShiftObject to allow to send the values in the POST request.
Displays a message if the shift is successfully dropped.