Time off requests modification
Created a card item for time off requests
Created an xml and a .kt file for time off requests.
Added functions to allow buttons to navigation to add another request, go back to the list, and show the list of requests after the user submits a new request again.
Made changes to the buttons on the home activity to allow proper navigation.
Added API in the apicalls.kt file
Created a new time off requests adapter.